+++ libraries/shared/src/MovingPercentile.h
@@ -18,19 +18,21 @@ class MovingPercentile {
 public:
     MovingPercentile(int numSamples, float percentile = 0.5f);
 
-    void updatePercentile(float sample);
-    float getValueAtPercentile() const { return _valueAtPercentile; }
+    // FIXME - this class is only currently used in calculating the clockSkew, which is a signed 64bit int, not a double
+    // I am somewhat tempted to make this a type sensitive template and/or swith to using qint64's internally
+    void updatePercentile(quint64 sample);
+    quint64 getValueAtPercentile() const { return _valueAtPercentile; }
 
 private:
     const int _numSamples;
     const float _percentile;
 
-    QList<float> _samplesSorted;
+    QList<quint64> _samplesSorted;
     QList<int> _sampleIds;      // incrementally assigned, is cyclic
     int _newSampleId;
 
     int _indexOfPercentile;
-    float _valueAtPercentile;
+    quint64 _valueAtPercentile;
 };
 
 #endif
